requirements.txt
accelerate==1.10.0
alembic==1.17.0
annotated-types==0.7.0
anthropic==0.62.0
anyio==4.10.0
asgiref==3.8.1
beautifulsoup4==4.13.4
bidict==0.23.1
brotli==1.2.0
cachetools==5.5.2
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.42.1
cmasher==1.9.2
cmocean==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
colorspacious==1.1.2
contourpy==1.3.0
-e git+https://github.com/InstituteforDiseaseModeling/covasim.git@148ea03c42b121dbc521098813a797a343edfffa#egg=covasim
coverage==7.8.2
cssselect2==0.8.0
cycler==0.12.1
dill==0.4.0
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-allauth>=0.60.0
doc8==2.0.0
docutils==0.21.2
EditorConfig==0.17.1
et_xmlfile==2.0.0
filelock==3.18.0
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.53.1
fsspec==2025.5.1
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
iniconfig==2.1.0
jellyfish==1.2.1
Jinja2==3.1.6
jiter==0.10.0
jsbeautifier==1.15.4
jsonobject==2.3.1
jsonpickle==4.1.1
kiwisolver==1.4.7
labelImg==1.8.6
labelme==5.8.1
lazy_loader==0.4
line_profiler==5.0.0
llvmlite==0.45.1
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
multiprocess==0.70.18
mysqlclient==2.2.7
narwhals==2.10.0
natsort==8.4.0
neo4j==5.28.1
networkx==3.5
numba==0.62.1
numpy==2.2.4
onnxruntime==1.22.0
openpyxl==3.1.5
optuna==4.5.0
osam==0.2.3
packaging==25.0
pandas==2.3.3
parestlib==0.4.0
patsy==1.0.2
pbr==6.1.0
peft==0.17.0
pillow==11.2.1
plotly==6.3.1
pluggy==1.6.0
pocketflow==0.0.3
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.5
pydantic_core==2.33.2
pydot==3.0.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.1.4
pyphen==0.17.2
PyPokerEngine==1.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
restructuredtext_lint==1.4.0
rich==14.2.0
rsa==4.9.1
rstcheck==6.2.5
rstcheck-core==1.2.2
rustworkx==0.15.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sciris==3.2.4
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.44
sqlparse==0.5.3
statsmodels==0.14.5
stevedore==5.3.0
#symengine==0.11.0
sympy==1.14.0
synthpops==1.10.5
termcolor==3.2.0
tifffile==2025.6.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.3.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0
win32_setctime==1.2.0
xlsxwriter==3.2.9
zopfli==0.2.3.post1
zstandard==0.25.0
